[gui] remove globals.dart
