+++ src/client/desktop_gui/lib/instance_actions_button.dart
@@ -3,7 +3,6 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import 'globals.dart';
 import 'grpc_client.dart';
 import 'instance_action.dart';
 import 'snackbar.dart';
