+++ src/client/desktop_gui/lib/vms_table.dart
@@ -4,7 +4,6 @@ import 'package:filesize/filesize.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import 'globals.dart';
 import 'grpc_client.dart';
 import 'instance_actions_button.dart';
 import 'vms_screen.dart';
